requirements.txt
diffusers
transformers
accelerate
opencv-contrib-python
controlnet_aux
torch
torchvision
pillow
numpy
gradio
